# specs/004-intent-specify/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AI 커리어 로드맵 플랫폼 API — 004 델타 (가치사슬 앞단·생태계·수익)
  version: 0.4.0
  description: >
    003 OpenAPI 위에 004 신규/확장 엔드포인트만 기술한 델타. 기존 인증(JWT)·에러 포맷·
    네트워크 표준(servers)을 계승한다. shared/types 는 base + 004 델타로 생성.
servers:
  - url: https://p16.sumzip.com/api/v1
  - url: http://localhost:9536/v1
...
paths:
...
  /students/me/profile-completeness:
    get:
      tags: [activities]
      summary: 활동·스펙 입력 충실도(문서·진단 전 안내용)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProfileCompleteness' }

  # ── G2 직무 맞춤 미션 ──
...
  /partners/banners/{id}/track:
    post:
      tags: [partners]
      summary: 배너 클릭/전환 집계(제휴 수수료 산정)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema: { type: object, properties: { event: { type: string, enum: [click, convert] } } }
      responses: { '204': { description: No Content } }

  # ── G4 파트너십 / G5 라이선스 (운영자/파트너) ──
  /admin/partners:
    post:
      tags: [partners]
      summary: 파트너 등록(운영자 수동, 역할 계정 발급)
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PartnerInput' }
      responses: { '201': { description: Created } }
...
  /admin/licenses:
    post:
      tags: [licenses]
      summary: 라이선스 계약 등록(대학 SaaS·기업 채용 수수료)
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: Created } }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    ProfileCompleteness:
      type: object
      properties:
        activities_count: { type: integer }
        credentials_count: { type: integer }
        ready_for_documents: { type: boolean }
        next_recommended_input: { type: string }
...
    PartnerInput:
      type: object
      required: [type, name]
      properties:
        type: { type: string, enum: [university, company, mentor_org, edu_platform, tech_partner] }
        name: { type: string }
        consent_scope: { type: string, enum: [none, stats, individual] }